+++ src/main/java/org/sbrubbles/genericcons/ParameterizedTypeNotFoundException.java
@@ -6,6 +6,8 @@ package org.sbrubbles.genericcons;
  * @author Humberto Anjos
  */
 public class ParameterizedTypeNotFoundException extends RuntimeException {
+  private static final long serialVersionUID = 42L;
+
   /**
    * Constructs a ParameterizedTypeNotFoundException with a default error message.
    */